pyproject.toml
[project]
name = "ANGANGBOT-QQapi"
version = "0.1.0"
description = "qqbot"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]

[tool.poetry]
name = "angangbot-qqapi"
version = "0.1.0"
description = ""
authors = ["angjustinl"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nb-cli = "^1.3.1"
nonebot2 = "^2.2.1"
nonebot-plugin-chatgpt-on-qq = "^1.6.4"
httpx = "^0.24"
requests = "^2.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
